moved sorting logic out of index page into utils
